+++ packages/swapper/src/swappers/ChainflipSwapper/endpoints.ts
@@ -18,7 +18,8 @@ import type {
 import { isExecutableTradeQuote, isExecutableTradeStep, isToken } from '../../utils'
 import { CHAINFLIP_BAAS_COMMISSION } from './constants'
 import type { ChainflipBaasSwapDepositAddress } from './models/ChainflipBaasSwapDepositAddress'
-import { getTradeQuote, getTradeRate } from './swapperApi/getTradeQuote'
+import { getTradeQuote } from './swapperApi/getTradeQuote'
+import { getTradeRate } from './swapperApi/getTradeRate'
 import type { ChainFlipStatus } from './types'
 import { chainflipService } from './utils/chainflipService'
 import { getLatestChainflipStatusMessage } from './utils/getLatestChainflipStatusMessage'
