+++ packages/swapper/src/swappers/ChainflipSwapper/swapperApi/getTradeQuote.ts
@@ -10,12 +10,10 @@ import { v4 as uuid } from 'uuid'
 import { getDefaultSlippageDecimalPercentageForSwapper } from '../../../constants'
 import type {
   CommonTradeQuoteInput,
-  GetTradeRateInput,
   GetUtxoTradeQuoteInput,
   SwapErrorRight,
   SwapperDeps,
   TradeQuote,
-  TradeRate,
 } from '../../../types'
 import {
   type GetEvmTradeQuoteInput,
@@ -40,7 +38,7 @@ import { getEvmTxFees } from '../utils/getEvmTxFees'
 import { getUtxoTxFees } from '../utils/getUtxoTxFees'
 import { getChainFlipIdFromAssetId, isSupportedAssetId, isSupportedChainId } from '../utils/helpers'
 
-const _getTradeQuote = async (
+export const _getTradeQuote = async (
   input: CommonTradeQuoteInput,
   deps: SwapperDeps,
 ): Promise<Result<TradeQuote[], SwapErrorRight>> => {
@@ -341,16 +339,6 @@ const _getTradeQuote = async (
   }
 
   return Ok(quotes)
-}
-
-// This isn't a mistake. A trade rate *is* a trade quote. Chainflip doesn't really have a notion of a trade quote,
-// they do have a notion of a "swap" (which we effectively only use to get the deposit address), which is irrelevant to the notion of quote vs. rate
-export const getTradeRate = async (
-  input: GetTradeRateInput,
-  deps: SwapperDeps,
-): Promise<Result<TradeRate[], SwapErrorRight>> => {
-  const rates = await _getTradeQuote(input as unknown as CommonTradeQuoteInput, deps)
-  return rates as Result<TradeRate[], SwapErrorRight>
 }
 
 export const getTradeQuote = async (
