+++ packages/swapper/src/swappers/ZrxSwapper/getZrxTradeRate/getZrxTradeRate.ts
@@ -1,0 +1,284 @@
+import type { ChainId } from '@shapeshiftoss/caip'
+import type { EvmChainAdapter } from '@shapeshiftoss/chain-adapters'
+import { evm } from '@shapeshiftoss/chain-adapters'
+import { PERMIT2_CONTRACT } from '@shapeshiftoss/contracts'
+import type { AssetsByIdPartial } from '@shapeshiftoss/types'
+import { bn, bnOrZero } from '@shapeshiftoss/utils'
+import type { Result } from '@sniptt/monads'
+import { Err, Ok } from '@sniptt/monads'
+import { v4 as uuid } from 'uuid'
+
+import { getDefaultSlippageDecimalPercentageForSwapper } from '../../../constants'
+import type {
+  GetEvmTradeRateInput,
+  SingleHopTradeRateSteps,
+  SwapErrorRight,
+  TradeRate,
+} from '../../../types'
+import { SwapperName, TradeQuoteError } from '../../../types'
+import { makeSwapErrorRight } from '../../../utils'
+import { isNativeEvmAsset } from '../../utils/helpers/helpers'
+import { fetchZrxPermit2Price, fetchZrxPrice } from '../utils/fetchFromZrx'
+import { isSupportedChainId } from '../utils/helpers/helpers'
+
+export function getZrxTradeRate(
+  input: GetEvmTradeRateInput,
+  assertGetEvmChainAdapter: (chainId: ChainId) => EvmChainAdapter,
+  isPermit2Enabled: boolean,
+  assetsById: AssetsByIdPartial,
+  zrxBaseUrl: string,
+): Promise<Result<TradeRate, SwapErrorRight>> {
+  if (!isPermit2Enabled) return _getZrxTradeRate(input, assertGetEvmChainAdapter, zrxBaseUrl)
+  return _getZrxPermit2TradeRate(input, assertGetEvmChainAdapter, assetsById, zrxBaseUrl)
+}
+
+async function _getZrxTradeRate(
+  input: GetEvmTradeRateInput,
+  assertGetEvmChainAdapter: (chainId: ChainId) => EvmChainAdapter,
+  zrxBaseUrl: string,
+): Promise<Result<TradeRate, SwapErrorRight>> {
+  const {
+    sellAsset,
+    buyAsset,
+    accountNumber,
+    receiveAddress,
+    affiliateBps,
+    potentialAffiliateBps,
+    chainId,
+    supportsEIP1559,
+    sellAmountIncludingProtocolFeesCryptoBaseUnit,
+  } = input
+
+  const slippageTolerancePercentageDecimal =
+    input.slippageTolerancePercentageDecimal ??
+    getDefaultSlippageDecimalPercentageForSwapper(SwapperName.Zrx)
+
+  const sellAssetChainId = sellAsset.chainId
+  const buyAssetChainId = buyAsset.chainId
+
+  if (!isSupportedChainId(sellAssetChainId)) {
+    return Err(
+      makeSwapErrorRight({
+        message: `unsupported chainId`,
+        code: TradeQuoteError.UnsupportedChain,
+        details: { chainId: sellAsset.chainId },
+      }),
+    )
+  }
+
+  if (!isSupportedChainId(buyAssetChainId)) {
+    return Err(
+      makeSwapErrorRight({
+        message: `unsupported chainId`,
+        code: TradeQuoteError.UnsupportedChain,
+        details: { chainId: sellAsset.chainId },
+      }),
+    )
+  }
+
+  if (sellAssetChainId !== buyAssetChainId) {
+    return Err(
+      makeSwapErrorRight({
+        message: `cross-chain not supported - both assets must be on chainId ${sellAsset.chainId}`,
+        code: TradeQuoteError.CrossChainNotSupported,
+        details: { buyAsset, sellAsset },
+      }),
+    )
+  }
+
+  const maybeZrxPriceResponse = await fetchZrxPrice({
+    buyAsset,
+    sellAsset,
+    sellAmountIncludingProtocolFeesCryptoBaseUnit,
+    // Cross-account not supported for ZRX
+    sellAddress: receiveAddress,
+    affiliateBps,
+    slippageTolerancePercentageDecimal,
+    zrxBaseUrl,
+  })
+
+  if (maybeZrxPriceResponse.isErr()) return Err(maybeZrxPriceResponse.unwrapErr())
+  const zrxPriceResponse = maybeZrxPriceResponse.unwrap()
+
+  const {
+    buyAmount: buyAmountAfterFeesCryptoBaseUnit,
+    grossBuyAmount: buyAmountBeforeFeesCryptoBaseUnit,
+    price,
+    allowanceTarget,
+    gas,
+    expectedSlippage,
+  } = zrxPriceResponse
+
+  const useSellAmount = !!sellAmountIncludingProtocolFeesCryptoBaseUnit
+  const rate = useSellAmount ? price : bn(1).div(price).toString()
+
+  // 0x approvals are cheaper than trades, but we don't have dynamic quote data for them.
+  // Instead, we use a hardcoded gasLimit estimate in place of the estimatedGas in the 0x quote response.
+  try {
+    const adapter = assertGetEvmChainAdapter(chainId)
+    const { average } = await adapter.getGasFeeData()
+    const networkFeeCryptoBaseUnit = evm.calcNetworkFeeCryptoBaseUnit({
+      ...average,
+      supportsEIP1559: Boolean(supportsEIP1559),
+      // add gas limit buffer to account for the fact we perform all of our validation on the trade quote estimations
+      // which are inaccurate and not what we use for the tx to broadcast
+      gasLimit: bnOrZero(gas).times(1.2).toFixed(),
+    })
+
+    return Ok({
+      id: uuid(),
+      accountNumber: undefined,
+      receiveAddress: undefined,
+      potentialAffiliateBps,
+      affiliateBps,
+      // Slippage protection is only provided for specific pairs.
+      // If slippage protection is not provided, assume a no slippage limit.
+      // If slippage protection is provided, return the limit instead of the estimated slippage.
+      // https://0x.org/docs/0x-swap-api/api-references/get-swap-v1-quote
+      slippageTolerancePercentageDecimal: expectedSlippage
+        ? slippageTolerancePercentageDecimal
+        : undefined,
+      rate,
+      steps: [
+        {
+          estimatedExecutionTimeMs: undefined,
+          allowanceContract: allowanceTarget,
+          buyAsset,
+          sellAsset,
+          accountNumber,
+          rate,
+          feeData: {
+            protocolFees: {},
+            networkFeeCryptoBaseUnit, // TODO(gomes): do we still want to handle L1 fee here for rates, since we leverage upstream fees calcs?
+          },
+          buyAmountBeforeFeesCryptoBaseUnit,
+          buyAmountAfterFeesCryptoBaseUnit,
+          sellAmountIncludingProtocolFeesCryptoBaseUnit,
+          source: SwapperName.Zrx,
+        },
+      ] as SingleHopTradeRateSteps,
+    })
+  } catch (err) {
+    return Err(
+      makeSwapErrorRight({
+        message: 'failed to get fee data',
+        cause: err,
+        code: TradeQuoteError.NetworkFeeEstimationFailed,
+      }),
+    )
+  }
+}
+
+async function _getZrxPermit2TradeRate(
+  input: GetEvmTradeRateInput,
+  _assertGetEvmChainAdapter: (chainId: ChainId) => EvmChainAdapter,
+  _assetsById: AssetsByIdPartial,
+  zrxBaseUrl: string,
+): Promise<Result<TradeRate, SwapErrorRight>> {
+  const {
+    sellAsset,
+    buyAsset,
+    accountNumber,
+    receiveAddress,
+    affiliateBps,
+    potentialAffiliateBps,
+    sellAmountIncludingProtocolFeesCryptoBaseUnit,
+  } = input
+
+  const slippageTolerancePercentageDecimal =
+    input.slippageTolerancePercentageDecimal ??
+    getDefaultSlippageDecimalPercentageForSwapper(SwapperName.Zrx)
+
+  const sellAssetChainId = sellAsset.chainId
+  const buyAssetChainId = buyAsset.chainId
+
+  if (!isSupportedChainId(sellAssetChainId)) {
+    return Err(
+      makeSwapErrorRight({
+        message: `unsupported chainId`,
+        code: TradeQuoteError.UnsupportedChain,
+        details: { chainId: sellAsset.chainId },
+      }),
+    )
+  }
+
+  if (!isSupportedChainId(buyAssetChainId)) {
+    return Err(
+      makeSwapErrorRight({
+        message: `unsupported chainId`,
+        code: TradeQuoteError.UnsupportedChain,
+        details: { chainId: sellAsset.chainId },
+      }),
+    )
+  }
+
+  if (sellAssetChainId !== buyAssetChainId) {
+    return Err(
+      makeSwapErrorRight({
+        message: `cross-chain not supported - both assets must be on chainId ${sellAsset.chainId}`,
+        code: TradeQuoteError.CrossChainNotSupported,
+        details: { buyAsset, sellAsset },
+      }),
+    )
+  }
+
+  const maybeZrxPriceResponse = await fetchZrxPermit2Price({
+    buyAsset,
+    sellAsset,
+    sellAmountIncludingProtocolFeesCryptoBaseUnit,
+    // Cross-account not supported for ZRX
+    sellAddress: receiveAddress,
+    affiliateBps,
+    slippageTolerancePercentageDecimal,
+    zrxBaseUrl,
+  })
+
+  if (maybeZrxPriceResponse.isErr()) return Err(maybeZrxPriceResponse.unwrapErr())
+  const zrxPriceResponse = maybeZrxPriceResponse.unwrap()
+
+  const {
+    buyAmount: buyAmountAfterFeesCryptoBaseUnit,
+    minBuyAmount: buyAmountBeforeFeesCryptoBaseUnit,
+    totalNetworkFee,
+  } = zrxPriceResponse
+
+  const rate = bnOrZero(buyAmountAfterFeesCryptoBaseUnit)
+    .div(sellAmountIncludingProtocolFeesCryptoBaseUnit)
+    .toString()
+
+  // 0x approvals are cheaper than trades, but we don't have dynamic quote data for them.
+  // Instead, we use a hardcoded gasLimit estimate in place of the estimatedGas in the 0x quote response.
+  const networkFeeCryptoBaseUnit = totalNetworkFee
+  return Ok({
+    id: uuid(),
+    accountNumber: undefined,
+    receiveAddress: undefined,
+    potentialAffiliateBps,
+    affiliateBps,
+    // Slippage protection is only provided for specific pairs.
+    // If slippage protection is not provided, assume a no slippage limit.
+    // If slippage protection is provided, return the limit instead of the estimated slippage.
+    // https://0x.org/docs/0x-swap-api/api-references/get-swap-v1-quote
+    slippageTolerancePercentageDecimal,
+    rate,
+    steps: [
+      {
+        estimatedExecutionTimeMs: undefined,
+        // We don't care about this - this is a rate, and if we really wanted to, we know the permit2 allowance target
+        allowanceContract: isNativeEvmAsset(sellAsset.assetId) ? undefined : PERMIT2_CONTRACT,
+        buyAsset,
+        sellAsset,
+        accountNumber,
+        rate,
+        feeData: {
+          protocolFees: {},
+          networkFeeCryptoBaseUnit, // L1 fee added inside of evm.calcNetworkFeeCryptoBaseUnit
+        },
+        buyAmountBeforeFeesCryptoBaseUnit,
+        buyAmountAfterFeesCryptoBaseUnit,
+        sellAmountIncludingProtocolFeesCryptoBaseUnit,
+        source: SwapperName.Zrx,
+      },
+    ] as unknown as SingleHopTradeRateSteps,
+  })
+}
